requirements.txt
anthropic==0.83.0
fastapi==0.131.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
httpx==0.28.1
pydantic==2.12.5
python-dotenv==1.2.1
uvicorn==0.41.0
